+++ libraries/base/changelog.md
@@ -40,6 +40,9 @@
 
   * Add `plusForeignPtr` to `Foreign.ForeignPtr`.
 
+  * Add `type family AppendSymbol (m :: Symbol) (n :: Symbol) :: Symbol` to `GHC.TypeLits`
+    (#12162)
+
 ## 4.9.0.0  *May 2016*
 
   * Bundled with GHC 8.0